buildings/pi/arduino-adapter/src/requirements.txt
paho-mqtt==1.6.1
colored==1.4.2
loguru==0.5.3
pysm
netifaces
pyserial